requirements.txt
fastapi~=0.115.8
asyncpg~=0.30.0
jinja2~=3.1.5
python-dotenv
pyjwt~=2.10.1
